Update
Changed the command to be encoded on some of the attacks (Sneakier).
Also changed the powershell persistence to bypass UAC if the user is a
local admin. The Bypass UAC will not work on Windows 10. (only 7/8/8.1)
Also added a check on the Excel macro so it is possible to run the excel
document 2 times without getting error messages.